+++ Source/Classes/BuildCompleter.ps1
@@ -2,10 +2,9 @@ using namespace System.Collections
 using namespace System.Collections.Generic
 using namespace System.Management.Automation
 
-# This class defines an attribute that allows the user the tab complete
-# build numbers for function parameters. For this completer to work the
-# users must have already provided the ProjectName parameter for the
-# function or set a default project.
+# This class defines an attribute that allows the user the tab complete  build numbers
+# for function parameters. For this completer to work the users must have already
+# provided the ProjectName parameter for the function, or set a default project.
 class BuildCompleter : IArgumentCompleter {
    [IEnumerable[CompletionResult]] CompleteArgument(
       [string] $CommandName,
@@ -30,8 +29,11 @@ class BuildCompleter : IArgumentCompleter {
       # list.
       if ($projectName) {
          foreach ($b in (Get-VSTeamBuild -ProjectName $projectName)) {
-            if ($b.buildNumber -like "$WordToComplete*") {
+            if ($b.buildNumber -like "$WordToComplete*" -and $b.buildNumber -notmatch '\S') {
                $results.Add([CompletionResult]::new($b.buildNumber))
+            }
+            elseif  ($b.buildNumber -like "*$WordToComplete*") {
+               $results.Add([CompletionResult]::new("'$($b.buildNumber.replace("'","''"))'", $b.buildNumber, 0, $b.buildNumber))
             }
          }
       }