+++ Source/Classes/ProjectValidateAttribute.ps1
@@ -8,12 +8,7 @@ class ProjectValidateAttribute : ValidateArgumentsAttribute {
       #Do not fail on null or empty, leave that to other validation conditions
       if ([string]::IsNullOrEmpty($arguments)) {return}
 
-      if (_hasProjectCacheExpired) {
-         [VSTeamProjectCache]::projects = _getProjects
-         [VSTeamProjectCache]::timestamp = (Get-Date).Minute
-      }
-
-      if (($null -ne [VSTeamProjectCache]::projects) -and (-not ($arguments -in [VSTeamProjectCache]::projects))) {
+      if (($null -ne [VSTeamProjectCache]::GetCurrent()) -and (-not ($arguments -in [VSTeamProjectCache]::projects))) {
          throw [ValidationMetadataException]::new(
             "'$arguments' is not a valid project. Valid projects are: '" +
             ([VSTeamProjectCache]::projects -join "', '") + "'")