+++ Source/Classes/VSTeamProjectCache.ps1
@@ -4,4 +4,14 @@
 class VSTeamProjectCache {
    static [int] $timestamp = -1
    static [object] $projects = $null
+   static [Void] Update () {
+      #Allow unit tests to mock returning the project list and testing freshness
+      [VSTeamProjectCache]::projects = _getProjects
+      [VSTeamProjectCache]::timestamp = (Get-Date).Minute
+      # "save current minute" refreshes on average after 30secs  but not after exact hours timeOfDayTotalMinutes might be a better base
+   }
+   static [object] GetCurrent () {
+      if (_hasProjectCacheExpired) { [VSTeamProjectCache]::Update() }
+      return ([VSTeamProjectCache]::projects)
+   }
 }