+++ Source/Public/Invoke-VSTeamRequest.ps1
@@ -1,9 +1,11 @@
 function Invoke-VSTeamRequest {
    [CmdletBinding()]
    param(
-      [string] $resource,
-
+      [ArgumentCompleter([AreaCompleter])]
       [string] $area,
+
+      [ArgumentCompleter([ResourceCompleter])]
+      [string] $resource,
 
       [string] $id,
 
@@ -32,26 +34,33 @@ function Invoke-VSTeamRequest {
       [object] $QueryString,
 
       [string] $Team,
-      
+
       [Parameter(Position = 0, ValueFromPipelineByPropertyName = $true)]
       [ProjectValidateAttribute()]
       [ArgumentCompleter([ProjectCompleter])]
       [string] $ProjectName,
-      
-      [switch] $UseProjectId,
-      
-      [switch] $NoProject)
 
+      [switch] $UseProjectId,
+
+      [switch] $NoProject,
+
+      [Alias('Value')]
+      [switch]$ExpandValue
+   )
    process {
       $params = $PSBoundParameters
 
       # We have to remove any extra parameters not used by Invoke-RestMethod
       $params.Remove('JSON') | Out-Null
+      $params.Remove('ExpandValue') | Out-Null
 
       $output = _callAPI @params
 
       if ($JSON.IsPresent) {
          $output | ConvertTo-Json -Depth 99
+      }
+      elseif ($ExpandValue) {
+         $output | Select-Object -ExpandProperty Value
       }
       else {
          $output