+++ Source/VSTeam.psm1
@@ -1,13 +1,5 @@
-# Dot sources all the files in Public, Private and Classes
-# Then reads the file names in Public and exports them
-# as the fucntions for the module.
-# To export an alias you have to manually add it to the
-# Export-ModuleMember below.
-# The module manifest is using wildcard exports for functions
-# and alias so you only have to name the files correctly.
-
-# Files are built using a script in the root folder
-. "$PSScriptRoot\vsteam.classes.ps1"
+# Dot source Public & Private functions
+# Files are built using a script in the root folder of the repository
 . "$PSScriptRoot\vsteam.functions.ps1"
 
 # Check to see if the user stored the default project in an environment variable